+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+# Unreleased
+
+-   Bump packages
+
 ## 3.1.0 - 2026-01-29
 
 -   Don't show the picker again after SDK version is installed