+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
+- Improve how versions are shown
 - Bump packages
 
 ## 3.1.0 - 2026-01-29