+++ CHANGELOG.md
@@ -4,8 +4,8 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
-- The _Quit Ren'Py_ button now closes the window without asking for a confirmation
-- Remove error message when initial socket connection cannot be made, and notify the user when waiting for a connection to be established
+- The _Quit Ren'Py_ button now closes the window without asking for confirmation in Ren'Py. This is achieved by overwriting `config.quit_action` when a session is active.
+- Remove error message when initial socket connection cannot be made, and instead notify the user when waiting for a connection to be established
 - Remove error message when game crashes (you already know that it crashed...)
 - Improve how versions are shown
 - There is now a way to return from the "Download SDK" menu